requirements.txt
fastapi
uvicorn
pandas
networkx
google-generativeai
pydantic
python-dotenv
